Fix IN with decimal in analyzer
