+++ src/Analyzer/SetUtils.cpp
@@ -70,7 +70,11 @@ Block createBlockFromCollection(const Collection & collection, const DataTypes &
     {
         if (columns_size == 1)
         {
-            auto field = convertFieldToType(value, *block_types[0]);
+            Field field;
+            bool is_conversion_ok = convertFieldToTypeStrict(value, *block_types[0], field);
+            if (!is_conversion_ok)
+                continue;
+
             bool need_insert_null = transform_null_in && block_types[0]->isNullable();
             if (!field.isNull() || need_insert_null)
                 columns[0]->insert(std::move(field));
@@ -98,7 +102,10 @@ Block createBlockFromCollection(const Collection & collection, const DataTypes &
         size_t i = 0;
         for (; i < tuple_size; ++i)
         {
-            tuple_values[i] = convertFieldToType(tuple[i], *block_types[i]);
+            bool is_conversion_ok = convertFieldToTypeStrict(tuple[i], *block_types[i], tuple_values[i]);
+            if (!is_conversion_ok)
+                break;
+
             bool need_insert_null = transform_null_in && block_types[i]->isNullable();
             if (tuple_values[i].isNull() && !need_insert_null)
                 break;
